+++ Changelog.md
@@ -6,6 +6,7 @@
 - for locations that have takeoff directions, the takeoff directions are now
   shown on the map as circle slices
 - automatically zoom to newly imported locations
+- added exporting CZML layers to .kmz files
 
 ## Version 1.9.0